@@ -13,23 +13,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.2.4",
-    "@angular/common": "17.2.4",
-    "@angular/compiler": "17.2.4",
-    "@angular/core": "17.2.4",
-    "@angular/forms": "17.2.4",
-    "@angular/platform-browser": "17.2.4",
-    "@angular/platform-browser-dynamic": "17.2.4",
-    "@angular/router": "17.2.4",
-    "angular-in-memory-web-api": "~0.17.0",
+    "@angular/animations": "19.1.6",
+    "@angular/common": "19.1.6",
+    "@angular/compiler": "19.1.6",
+    "@angular/core": "19.1.6",
+    "@angular/forms": "19.1.6",
+    "@angular/platform-browser": "19.1.6",
+    "@angular/platform-browser-dynamic": "19.1.6",
+    "@angular/router": "19.1.6",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.0"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.2.3",
-    "@angular/cli": "17.2.3",
-    "@angular/compiler-cli": "17.2.4",
+    "@angular-devkit/build-angular": "19.1.7",
+    "@angular/cli": "19.1.7",
+    "@angular/compiler-cli": "19.1.6",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^16.11.35",
     "copyfiles": "^2.4.1",
@@ -43,6 +42,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "protractor": "~7.0.0",
     "ts-node": "~10.9.0",
-    "typescript": "~4.9.3"
+    "typescript": "~5.5.4"
   }
 }
