@@ -43,8 +43,6 @@ import {
   NumberSymbol,
   PLATFORM_BROWSER_ID,
   PLATFORM_SERVER_ID,
-  PLATFORM_WORKER_APP_ID,
-  PLATFORM_WORKER_UI_ID,
   PRECONNECT_CHECK_BLOCKLIST,
   PathLocationStrategy,
   PercentPipe,
@@ -87,8 +85,6 @@ import {
   getNumberOfCurrencyDigits,
   isPlatformBrowser,
   isPlatformServer,
-  isPlatformWorkerApp,
-  isPlatformWorkerUi,
   normalizeQueryParams,
   parseCookieValue,
   provideCloudflareLoader,
@@ -98,10 +94,10 @@ import {
   provideNetlifyLoader,
   registerLocaleData,
   setRootDomAdapter
-} from "./chunk-AE3ITOJK.js";
+} from "./chunk-6OORFGZS.js";
 import {
   IMAGE_CONFIG
-} from "./chunk-47TOL4VA.js";
+} from "./chunk-KQVTM4CV.js";
 export {
   APP_BASE_HREF,
   AsyncPipe,
@@ -185,8 +181,6 @@ export {
   getNumberOfCurrencyDigits,
   isPlatformBrowser,
   isPlatformServer,
-  isPlatformWorkerApp,
-  isPlatformWorkerUi,
   provideCloudflareLoader,
   provideCloudinaryLoader,
   provideImageKitLoader,
@@ -197,12 +191,9 @@ export {
   NullViewportScroller as ɵNullViewportScroller,
   PLATFORM_BROWSER_ID as ɵPLATFORM_BROWSER_ID,
   PLATFORM_SERVER_ID as ɵPLATFORM_SERVER_ID,
-  PLATFORM_WORKER_APP_ID as ɵPLATFORM_WORKER_APP_ID,
-  PLATFORM_WORKER_UI_ID as ɵPLATFORM_WORKER_UI_ID,
   PlatformNavigation as ɵPlatformNavigation,
   getDOM as ɵgetDOM,
   normalizeQueryParams as ɵnormalizeQueryParams,
   parseCookieValue as ɵparseCookieValue,
   setRootDomAdapter as ɵsetRootDomAdapter
 };
-//# sourceMappingURL=@angular_common.js.map
