@@ -2,6 +2,7 @@ import {
   FetchBackend,
   HTTP_INTERCEPTORS,
   HTTP_ROOT_INTERCEPTOR_FNS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
   HttpBackend,
   HttpClient,
   HttpClientJsonpModule,
@@ -26,7 +27,7 @@ import {
   HttpXsrfTokenExtractor,
   JsonpClientBackend,
   JsonpInterceptor,
-  PRIMARY_HTTP_BACKEND,
+  REQUESTS_CONTRIBUTE_TO_STABILITY,
   provideHttpClient,
   withFetch,
   withHttpTransferCache,
@@ -36,12 +37,13 @@ import {
   withNoXsrfProtection,
   withRequestsMadeViaParent,
   withXsrfConfiguration
-} from "./chunk-GQGXAMSG.js";
-import "./chunk-AE3ITOJK.js";
-import "./chunk-47TOL4VA.js";
+} from "./chunk-N7BYCKIZ.js";
+import "./chunk-6OORFGZS.js";
+import "./chunk-KQVTM4CV.js";
 export {
   FetchBackend,
   HTTP_INTERCEPTORS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
   HttpBackend,
   HttpClient,
   HttpClientJsonpModule,
@@ -76,7 +78,6 @@ export {
   HTTP_ROOT_INTERCEPTOR_FNS as ɵHTTP_ROOT_INTERCEPTOR_FNS,
   HttpInterceptorHandler as ɵHttpInterceptingHandler,
   HttpInterceptorHandler as ɵHttpInterceptorHandler,
-  PRIMARY_HTTP_BACKEND as ɵPRIMARY_HTTP_BACKEND,
+  REQUESTS_CONTRIBUTE_TO_STABILITY as ɵREQUESTS_CONTRIBUTE_TO_STABILITY,
   withHttpTransferCache as ɵwithHttpTransferCache
 };
-//# sourceMappingURL=@angular_common_http.js.map
