@@ -4,7 +4,7 @@ import {
   APP_BOOTSTRAP_LISTENER,
   APP_ID,
   APP_INITIALIZER,
-  AfterRenderEventManager,
+  AfterRenderManager,
   AfterRenderPhase,
   ApplicationInitStatus,
   ApplicationModule,
@@ -15,6 +15,7 @@ import {
   CSP_NONCE,
   CUSTOM_ELEMENTS_SCHEMA,
   ChangeDetectionScheduler,
+  ChangeDetectionSchedulerImpl,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Compiler,
@@ -40,6 +41,7 @@ import {
   DeferBlockState,
   DestroyRef,
   Directive,
+  ENABLE_ROOT_COMPONENT_BOOTSTRAP,
   ENVIRONMENT_INITIALIZER,
   EffectScheduler,
   ElementRef,
@@ -48,27 +50,31 @@ import {
   ErrorHandler,
   EventEmitter,
   FactoryTarget,
+  HOST_TAG_NAME,
   Host,
+  HostAttributeToken,
   HostBinding,
   HostListener,
   IMAGE_CONFIG,
   IMAGE_CONFIG_DEFAULTS,
-  INJECTOR,
+  INJECTOR$1,
   INJECTOR_SCOPE,
+  INTERNAL_APPLICATION_ERROR_HANDLER,
   IS_HYDRATION_DOM_REUSE_ENABLED,
+  IS_INCREMENTAL_HYDRATION_ENABLED,
   Inject,
   InjectFlags,
   Injectable,
   InjectionToken,
   Injector,
   Input,
-  InputFlags,
   IterableDiffers,
+  JSACTION_EVENT_CONTRACT,
   KeyValueDiffers,
   LContext,
   LOCALE_ID,
-  LifecycleHooksFeature,
   LocaleDataIndex,
+  MicrotaskEffectScheduler,
   MissingTranslationStrategy,
   ModuleWithComponentFactories,
   NG_COMP_DEF,
@@ -91,19 +97,28 @@ import {
   NoopNgZone,
   Optional,
   Output,
+  OutputEmitterRef,
   PACKAGE_ROOT_URL,
+  PERFORMANCE_MARK_PREFIX,
   PLATFORM_ID,
   PLATFORM_INITIALIZER,
+  PROVIDED_NG_ZONE,
   PendingTasks,
+  PendingTasksInternal,
   Pipe,
   PlatformRef,
   Query,
   QueryList,
+  REQUEST,
+  REQUEST_CONTEXT,
+  RESPONSE_INIT,
   ReflectionCapabilities,
   Renderer2,
   RendererFactory2,
   RendererStyleFlags2,
+  ResourceStatus,
   RuntimeError,
+  SIGNAL,
   SSR_CONTENT_INTEGRITY_MARKER,
   Sanitizer,
   SecurityContext,
@@ -117,6 +132,8 @@ import {
   TemplateRef,
   Testability,
   TestabilityRegistry,
+  TracingAction,
+  TracingService,
   TransferState,
   Type,
   USE_RUNTIME_DEPS_TRACKER_FOR_JIT,
@@ -125,15 +142,17 @@ import {
   ViewChild,
   ViewChildren,
   ViewContainerRef,
-  ViewEncapsulation$1,
+  ViewEncapsulation,
   ViewRef,
   ViewRef$1,
   XSS_SECURITY_URL,
+  ZONELESS_ENABLED,
   _global,
   _sanitizeHtml,
   _sanitizeUrl,
   afterNextRender,
   afterRender,
+  afterRenderEffect,
   allowSanitizationBypassAndThrow,
   annotateForHydration,
   asNativeElements,
@@ -169,19 +188,22 @@ import {
   defineInjectable,
   depsTracker,
   destroyPlatform,
+  detectChangesInViewIfRequired,
   devModeEqual,
+  disableProfiling,
   effect,
   enableProdMode,
+  enableProfiling,
   findLocaleData,
   flushModuleScopingQueueAsMuchAsPossible,
   formatRuntimeError,
   forwardRef,
   generateStandaloneInDeclarationsError,
   getAsyncClassMetadataFn,
+  getClosestComponentName,
   getDebugNode,
-  getDeferBlocks,
+  getDeferBlocks$1,
   getDirectives,
-  getEnsureDirtyViewsAreAlwaysReachable,
   getHostElement,
   getInjectableDef,
   getLContext,
@@ -189,14 +211,15 @@ import {
   getLocalePluralCase,
   getModuleFactory,
   getNgModuleById,
+  getOutputDestroyRef,
   getPlatform,
   getSanitizationBypassType,
   importProvidersFrom,
   inject,
   injectChangeDetectorRef,
   input,
-  internalAfterNextRender,
   internalCreateApplication,
+  internalProvideZoneChangeDetection,
   isBoundToModule,
   isComponentDefPendingResolution,
   isDevMode,
@@ -207,9 +230,11 @@ import {
   isSignal,
   isStandalone,
   isSubscribable,
+  linkedSignal,
   makeEnvironmentProviders,
   makeStateKey,
   mergeApplicationConfig,
+  microtaskEffect,
   model,
   noSideEffects,
   numberAttribute,
@@ -217,9 +242,13 @@ import {
   patchComponentDefWithScope,
   performanceMarkFeature,
   platformCore,
+  provideAppInitializer,
+  provideEnvironmentInitializer,
+  provideExperimentalCheckNoChangesForDebug,
+  provideExperimentalZonelessChangeDetection,
+  providePlatformInitializer,
   provideZoneChangeDetection,
-  provideZonelessChangeDetection,
-  queueStateUpdate,
+  publishExternalGlobalUtil,
   readHydrationInfo,
   reflectComponentType,
   registerLocaleData,
@@ -229,6 +258,7 @@ import {
   resetJitOptions,
   resolveComponentResources,
   resolveForwardRef,
+  resource,
   restoreComponentResolutionQueue,
   runInInjectionContext,
   setAllowDuplicateNgModuleIdsForTest,
@@ -237,11 +267,12 @@ import {
   setClassMetadataAsync,
   setCurrentInjector,
   setDocument,
-  setEnsureDirtyViewsAreAlwaysReachable,
   setInjectorProfilerContext,
   setLocaleId,
   setTestabilityGetter,
   signal,
+  startMeasuring,
+  stopMeasuring,
   store,
   stringify,
   transitiveScopesFor,
@@ -252,8 +283,10 @@ import {
   unwrapSafeValue,
   viewChild,
   viewChildren,
-  whenStable,
   withDomHydration,
+  withEventReplay,
+  withI18nSupport,
+  withIncrementalHydration,
   ɵINPUT_SIGNAL_BRAND_WRITE_TYPE,
   ɵgetUnknownElementStrictMode,
   ɵgetUnknownPropertyStrictMode,
@@ -262,13 +295,14 @@ import {
   ɵsetUnknownPropertyStrictMode,
   ɵunwrapWritableSignal,
   ɵɵCopyDefinitionFeature,
+  ɵɵExternalStylesFeature,
   ɵɵHostDirectivesFeature,
   ɵɵInheritDefinitionFeature,
   ɵɵInputTransformsFeature,
   ɵɵNgOnChangesFeature,
   ɵɵProvidersFeature,
-  ɵɵStandaloneFeature,
   ɵɵadvance,
+  ɵɵattachSourceLocations,
   ɵɵattribute,
   ɵɵattributeInterpolate1,
   ɵɵattributeInterpolate2,
@@ -294,8 +328,17 @@ import {
   ɵɵconditional,
   ɵɵcontentQuery,
   ɵɵcontentQuerySignal,
+  ɵɵdeclareLet,
   ɵɵdefer,
   ɵɵdeferEnableTimerScheduling,
+  ɵɵdeferHydrateNever,
+  ɵɵdeferHydrateOnHover,
+  ɵɵdeferHydrateOnIdle,
+  ɵɵdeferHydrateOnImmediate,
+  ɵɵdeferHydrateOnInteraction,
+  ɵɵdeferHydrateOnTimer,
+  ɵɵdeferHydrateOnViewport,
+  ɵɵdeferHydrateWhen,
   ɵɵdeferOnHover,
   ɵɵdeferOnIdle,
   ɵɵdeferOnImmediate,
@@ -347,6 +390,7 @@ import {
   ɵɵnamespaceSVG,
   ɵɵnextContext,
   ɵɵngDeclareClassMetadata,
+  ɵɵngDeclareClassMetadataAsync,
   ɵɵngDeclareComponent,
   ɵɵngDeclareDirective,
   ɵɵngDeclareFactory,
@@ -385,11 +429,13 @@ import {
   ɵɵpureFunctionV,
   ɵɵqueryAdvance,
   ɵɵqueryRefresh,
+  ɵɵreadContextLet,
   ɵɵreference,
   ɵɵrepeater,
   ɵɵrepeaterCreate,
   ɵɵrepeaterTrackByIdentity,
   ɵɵrepeaterTrackByIndex,
+  ɵɵreplaceMetadata,
   ɵɵresetView,
   ɵɵresolveBody,
   ɵɵresolveDocument,
@@ -403,6 +449,7 @@ import {
   ɵɵsanitizeUrlOrResourceUrl,
   ɵɵsetComponentScope,
   ɵɵsetNgModuleScope,
+  ɵɵstoreLet,
   ɵɵstyleMap,
   ɵɵstyleMapInterpolate1,
   ɵɵstyleMapInterpolate2,
@@ -446,7 +493,7 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-47TOL4VA.js";
+} from "./chunk-KQVTM4CV.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
@@ -483,10 +530,12 @@ export {
   EnvironmentInjector,
   ErrorHandler,
   EventEmitter,
+  HOST_TAG_NAME,
   Host,
+  HostAttributeToken,
   HostBinding,
   HostListener,
-  INJECTOR,
+  INJECTOR$1 as INJECTOR,
   Inject,
   InjectFlags,
   Injectable,
@@ -506,16 +555,22 @@ export {
   NgZone,
   Optional,
   Output,
+  OutputEmitterRef,
   PACKAGE_ROOT_URL,
   PLATFORM_ID,
   PLATFORM_INITIALIZER,
+  PendingTasks,
   Pipe,
   PlatformRef,
   Query,
   QueryList,
+  REQUEST,
+  REQUEST_CONTEXT,
+  RESPONSE_INIT,
   Renderer2,
   RendererFactory2,
   RendererStyleFlags2,
+  ResourceStatus,
   Sanitizer,
   SecurityContext,
   Self,
@@ -533,10 +588,11 @@ export {
   ViewChild,
   ViewChildren,
   ViewContainerRef,
-  ViewEncapsulation$1 as ViewEncapsulation,
+  ViewEncapsulation,
   ViewRef,
   afterNextRender,
   afterRender,
+  afterRenderEffect,
   asNativeElements,
   assertInInjectionContext,
   assertNotInReactiveContext,
@@ -566,15 +622,23 @@ export {
   isDevMode,
   isSignal,
   isStandalone,
+  linkedSignal,
   makeEnvironmentProviders,
   makeStateKey,
   mergeApplicationConfig,
   model,
   numberAttribute,
+  output,
   platformCore,
+  provideAppInitializer,
+  provideEnvironmentInitializer,
+  provideExperimentalCheckNoChangesForDebug,
+  provideExperimentalZonelessChangeDetection,
+  providePlatformInitializer,
   provideZoneChangeDetection,
   reflectComponentType,
   resolveForwardRef,
+  resource,
   runInInjectionContext,
   setTestabilityGetter,
   signal,
@@ -582,9 +646,10 @@ export {
   viewChild,
   viewChildren,
   ALLOW_MULTIPLE_PLATFORMS as ɵALLOW_MULTIPLE_PLATFORMS,
-  AfterRenderEventManager as ɵAfterRenderEventManager,
+  AfterRenderManager as ɵAfterRenderManager,
   CONTAINER_HEADER_OFFSET as ɵCONTAINER_HEADER_OFFSET,
   ChangeDetectionScheduler as ɵChangeDetectionScheduler,
+  ChangeDetectionSchedulerImpl as ɵChangeDetectionSchedulerImpl,
   ComponentFactory$1 as ɵComponentFactory,
   Console as ɵConsole,
   DEFAULT_LOCALE_ID as ɵDEFAULT_LOCALE_ID,
@@ -592,15 +657,19 @@ export {
   DEFER_BLOCK_DEPENDENCY_INTERCEPTOR as ɵDEFER_BLOCK_DEPENDENCY_INTERCEPTOR,
   DeferBlockBehavior as ɵDeferBlockBehavior,
   DeferBlockState as ɵDeferBlockState,
+  ENABLE_ROOT_COMPONENT_BOOTSTRAP as ɵENABLE_ROOT_COMPONENT_BOOTSTRAP,
   EffectScheduler as ɵEffectScheduler,
   IMAGE_CONFIG as ɵIMAGE_CONFIG,
   IMAGE_CONFIG_DEFAULTS as ɵIMAGE_CONFIG_DEFAULTS,
   INJECTOR_SCOPE as ɵINJECTOR_SCOPE,
   ɵINPUT_SIGNAL_BRAND_WRITE_TYPE,
+  INTERNAL_APPLICATION_ERROR_HANDLER as ɵINTERNAL_APPLICATION_ERROR_HANDLER,
   IS_HYDRATION_DOM_REUSE_ENABLED as ɵIS_HYDRATION_DOM_REUSE_ENABLED,
+  IS_INCREMENTAL_HYDRATION_ENABLED as ɵIS_INCREMENTAL_HYDRATION_ENABLED,
+  JSACTION_EVENT_CONTRACT as ɵJSACTION_EVENT_CONTRACT,
   LContext as ɵLContext,
-  LifecycleHooksFeature as ɵLifecycleHooksFeature,
   LocaleDataIndex as ɵLocaleDataIndex,
+  MicrotaskEffectScheduler as ɵMicrotaskEffectScheduler,
   NG_COMP_DEF as ɵNG_COMP_DEF,
   NG_DIR_DEF as ɵNG_DIR_DEF,
   NG_ELEMENT_ID as ɵNG_ELEMENT_ID,
@@ -612,18 +681,24 @@ export {
   NO_CHANGE as ɵNO_CHANGE,
   NgModuleFactory as ɵNgModuleFactory,
   NoopNgZone as ɵNoopNgZone,
-  PendingTasks as ɵPendingTasks,
+  PERFORMANCE_MARK_PREFIX as ɵPERFORMANCE_MARK_PREFIX,
+  PROVIDED_NG_ZONE as ɵPROVIDED_NG_ZONE,
+  PendingTasksInternal as ɵPendingTasksInternal,
   ReflectionCapabilities as ɵReflectionCapabilities,
   ComponentFactory as ɵRender3ComponentFactory,
   ComponentRef as ɵRender3ComponentRef,
   NgModuleRef as ɵRender3NgModuleRef,
   RuntimeError as ɵRuntimeError,
+  SIGNAL as ɵSIGNAL,
   SSR_CONTENT_INTEGRITY_MARKER as ɵSSR_CONTENT_INTEGRITY_MARKER,
   TESTABILITY as ɵTESTABILITY,
   TESTABILITY_GETTER as ɵTESTABILITY_GETTER,
+  TracingAction as ɵTracingAction,
+  TracingService as ɵTracingService,
   USE_RUNTIME_DEPS_TRACKER_FOR_JIT as ɵUSE_RUNTIME_DEPS_TRACKER_FOR_JIT,
   ViewRef$1 as ɵViewRef,
   XSS_SECURITY_URL as ɵXSS_SECURITY_URL,
+  ZONELESS_ENABLED as ɵZONELESS_ENABLED,
   _sanitizeHtml as ɵ_sanitizeHtml,
   _sanitizeUrl as ɵ_sanitizeUrl,
   allowSanitizationBypassAndThrow as ɵallowSanitizationBypassAndThrow,
@@ -645,28 +720,32 @@ export {
   defaultIterableDiffers as ɵdefaultIterableDiffers,
   defaultKeyValueDiffers as ɵdefaultKeyValueDiffers,
   depsTracker as ɵdepsTracker,
+  detectChangesInViewIfRequired as ɵdetectChangesInViewIfRequired,
   devModeEqual as ɵdevModeEqual,
+  disableProfiling as ɵdisableProfiling,
+  enableProfiling as ɵenableProfiling,
   findLocaleData as ɵfindLocaleData,
   flushModuleScopingQueueAsMuchAsPossible as ɵflushModuleScopingQueueAsMuchAsPossible,
   formatRuntimeError as ɵformatRuntimeError,
   generateStandaloneInDeclarationsError as ɵgenerateStandaloneInDeclarationsError,
   getAsyncClassMetadataFn as ɵgetAsyncClassMetadataFn,
+  getClosestComponentName as ɵgetClosestComponentName,
   getDebugNode as ɵgetDebugNode,
-  getDeferBlocks as ɵgetDeferBlocks,
+  getDeferBlocks$1 as ɵgetDeferBlocks,
   getDirectives as ɵgetDirectives,
-  getEnsureDirtyViewsAreAlwaysReachable as ɵgetEnsureDirtyViewsAreAlwaysReachable,
   getHostElement as ɵgetHostElement,
   getInjectableDef as ɵgetInjectableDef,
   getLContext as ɵgetLContext,
   getLocaleCurrencyCode as ɵgetLocaleCurrencyCode,
   getLocalePluralCase as ɵgetLocalePluralCase,
+  getOutputDestroyRef as ɵgetOutputDestroyRef,
   getSanitizationBypassType as ɵgetSanitizationBypassType,
   ɵgetUnknownElementStrictMode,
   ɵgetUnknownPropertyStrictMode,
   _global as ɵglobal,
   injectChangeDetectorRef as ɵinjectChangeDetectorRef,
-  internalAfterNextRender as ɵinternalAfterNextRender,
   internalCreateApplication as ɵinternalCreateApplication,
+  internalProvideZoneChangeDetection as ɵinternalProvideZoneChangeDetection,
   isBoundToModule as ɵisBoundToModule,
   isComponentDefPendingResolution as ɵisComponentDefPendingResolution,
   isEnvironmentProviders as ɵisEnvironmentProviders,
@@ -674,12 +753,11 @@ export {
   isNgModule as ɵisNgModule,
   isPromise as ɵisPromise,
   isSubscribable as ɵisSubscribable,
+  microtaskEffect as ɵmicrotaskEffect,
   noSideEffects as ɵnoSideEffects,
-  output as ɵoutput,
   patchComponentDefWithScope as ɵpatchComponentDefWithScope,
   performanceMarkFeature as ɵperformanceMarkFeature,
-  provideZonelessChangeDetection as ɵprovideZonelessChangeDetection,
-  queueStateUpdate as ɵqueueStateUpdate,
+  publishExternalGlobalUtil as ɵpublishExternalGlobalUtil,
   readHydrationInfo as ɵreadHydrationInfo,
   registerLocaleData as ɵregisterLocaleData,
   renderDeferBlockState as ɵrenderDeferBlockState,
@@ -694,11 +772,12 @@ export {
   setClassMetadataAsync as ɵsetClassMetadataAsync,
   setCurrentInjector as ɵsetCurrentInjector,
   setDocument as ɵsetDocument,
-  setEnsureDirtyViewsAreAlwaysReachable as ɵsetEnsureDirtyViewsAreAlwaysReachable,
   setInjectorProfilerContext as ɵsetInjectorProfilerContext,
   setLocaleId as ɵsetLocaleId,
   ɵsetUnknownElementStrictMode,
   ɵsetUnknownPropertyStrictMode,
+  startMeasuring as ɵstartMeasuring,
+  stopMeasuring as ɵstopMeasuring,
   store as ɵstore,
   stringify as ɵstringify,
   transitiveScopesFor as ɵtransitiveScopesFor,
@@ -707,18 +786,20 @@ export {
   unregisterAllLocaleData as ɵunregisterLocaleData,
   unwrapSafeValue as ɵunwrapSafeValue,
   ɵunwrapWritableSignal,
-  whenStable as ɵwhenStable,
   withDomHydration as ɵwithDomHydration,
+  withEventReplay as ɵwithEventReplay,
+  withI18nSupport as ɵwithI18nSupport,
+  withIncrementalHydration as ɵwithIncrementalHydration,
   ɵɵCopyDefinitionFeature,
+  ɵɵExternalStylesFeature,
   FactoryTarget as ɵɵFactoryTarget,
   ɵɵHostDirectivesFeature,
   ɵɵInheritDefinitionFeature,
-  InputFlags as ɵɵInputFlags,
   ɵɵInputTransformsFeature,
   ɵɵNgOnChangesFeature,
   ɵɵProvidersFeature,
-  ɵɵStandaloneFeature,
   ɵɵadvance,
+  ɵɵattachSourceLocations,
   ɵɵattribute,
   ɵɵattributeInterpolate1,
   ɵɵattributeInterpolate2,
@@ -744,8 +825,17 @@ export {
   ɵɵconditional,
   ɵɵcontentQuery,
   ɵɵcontentQuerySignal,
+  ɵɵdeclareLet,
   ɵɵdefer,
   ɵɵdeferEnableTimerScheduling,
+  ɵɵdeferHydrateNever,
+  ɵɵdeferHydrateOnHover,
+  ɵɵdeferHydrateOnIdle,
+  ɵɵdeferHydrateOnImmediate,
+  ɵɵdeferHydrateOnInteraction,
+  ɵɵdeferHydrateOnTimer,
+  ɵɵdeferHydrateOnViewport,
+  ɵɵdeferHydrateWhen,
   ɵɵdeferOnHover,
   ɵɵdeferOnIdle,
   ɵɵdeferOnImmediate,
@@ -797,6 +887,7 @@ export {
   ɵɵnamespaceSVG,
   ɵɵnextContext,
   ɵɵngDeclareClassMetadata,
+  ɵɵngDeclareClassMetadataAsync,
   ɵɵngDeclareComponent,
   ɵɵngDeclareDirective,
   ɵɵngDeclareFactory,
@@ -835,12 +926,14 @@ export {
   ɵɵpureFunctionV,
   ɵɵqueryAdvance,
   ɵɵqueryRefresh,
+  ɵɵreadContextLet,
   ɵɵreference,
   registerNgModuleType as ɵɵregisterNgModuleType,
   ɵɵrepeater,
   ɵɵrepeaterCreate,
   ɵɵrepeaterTrackByIdentity,
   ɵɵrepeaterTrackByIndex,
+  ɵɵreplaceMetadata,
   ɵɵresetView,
   ɵɵresolveBody,
   ɵɵresolveDocument,
@@ -854,6 +947,7 @@ export {
   ɵɵsanitizeUrlOrResourceUrl,
   ɵɵsetComponentScope,
   ɵɵsetNgModuleScope,
+  ɵɵstoreLet,
   ɵɵstyleMap,
   ɵɵstyleMapInterpolate1,
   ɵɵstyleMapInterpolate2,
@@ -898,4 +992,3 @@ export {
   ɵɵviewQuery,
   ɵɵviewQuerySignal
 };
-//# sourceMappingURL=@angular_core.js.map
