@@ -1,18 +1,18 @@
-import { HttpClient } from '@angular/common/http';
-import { Product } from './products';
 import { Injectable } from '@angular/core';
+import { Product } from './products';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CartService {
+items : Product[] = [];
 
-  items: Product[] = [];
-  // items: {product: Product, quantity: number }[] = [];
+  constructor(
+    private http: HttpClient
+  ) { }
 
-  constructor(private http: HttpClient) { }
-
-  addToCart(product: Product) {
+  addToCart( product: Product) {
     this.items.push(product);
   }
 
@@ -26,6 +26,7 @@ export class CartService {
   }
 
   getShippingPrices() {
-    return this.http.get<{type: string, price: number}[]>('/assets/shipping.json');
+    return this.http.get<{type: string, price: number}[]>('assets/shipping.json')
   }
+
 }
