@@ -1,24 +1,29 @@
 import { Component } from '@angular/core';
+import { CartService } from '../cart.service';
 import { FormBuilder } from '@angular/forms';
 
-import { CartService } from '../cart.service';
-
 @Component({
-  selector: 'app-cart',
-  templateUrl: './cart.component.html',
-  styleUrl: './cart.component.css'
+    selector: 'app-cart',
+    templateUrl: './cart.component.html',
+    styleUrl: './cart.component.css',
+    standalone: false
 })
 export class CartComponent {
 
   items = this.cartService.getItems();
-  checkoutForm = this.formBuilder.group({ name: '', address: '' });
-  
-  constructor(private cartService: CartService, private formBuilder: FormBuilder) { }
+  checkoutForm = this.formBuilder.group({
+    name: '',
+    address: '' 
+  });
 
-  onSubmit(): void {
-    // Process checkout data here
+  constructor (
+    private cartService : CartService,
+    private formBuilder : FormBuilder
+  ){}
+
+  onSubmit() : void  {
     this.items = this.cartService.clearCart();
-    console.warn('Your order has been submitted', this.checkoutForm.value);
+    console.warn(' pedido enviado', this.checkoutForm.value);
     this.checkoutForm.reset();
   }
 }
