@@ -1,13 +1,16 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, Output, EventEmitter,} from '@angular/core';
 import { Product } from '../products';
 
 @Component({
-  selector: 'app-product-alerts',
-  templateUrl: './product-alerts.component.html',
-  styleUrl: './product-alerts.component.css'
+    selector: 'app-product-alerts',
+    templateUrl: './product-alerts.component.html',
+    styleUrl: './product-alerts.component.css',
+    standalone: false
 })
 export class ProductAlertsComponent {
 
   @Input() product: Product | undefined;
-  @Output() notify = new EventEmitter();
+  
+  @Output() notificacion = new EventEmitter();
+  
 }
