@@ -1,31 +1,32 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-
 import { Product, products } from '../products';
 import { CartService } from '../cart.service';
 
-@Component({
-  selector: 'app-product-details',
-  templateUrl: './product-details.component.html',
-  styleUrl: './product-details.component.css'
-})
-export class ProductDetailsComponent implements OnInit {
-  
-  product: Product | undefined;
 
-  constructor(private route: ActivatedRoute, private cartService: CartService) { }
+@Component({
+    selector: 'app-product-details',
+    templateUrl: './product-details.component.html',
+    styleUrl: './product-details.component.css',
+    standalone: false
+})
+export class ProductDetailsComponent {
+
+  product: Product | undefined;
+  constructor (
+     private route: ActivatedRoute,
+    private cartService: CartService
+    ) {}
 
   ngOnInit() {
-    // First get the product id from the current route.
     const routeParams = this.route.snapshot.paramMap;
     const productIdFromRoute = Number(routeParams.get('productId'));
-  
-    // Find the product that correspond with the id provided in route.
-    this.product = products.find(product => product.id === productIdFromRoute);
-  }
 
-  addToCart(product: Product) {
+    this.product = products.find(product => product.id === productIdFromRoute);
+  
+  }
+  addToCart(product : Product) {
     this.cartService.addToCart(product);
-    window.alert('Your product has been added to the cart!');
+    window.alert('Producto añadido al carro');
   }
 }
