@@ -3,19 +3,20 @@ import { Component } from '@angular/core';
 import { products } from '../products';
 
 @Component({
-  selector: 'app-product-list',
-  templateUrl: './product-list.component.html',
-  styleUrls: ['./product-list.component.css']
+    selector: 'app-product-list',
+    templateUrl: './product-list.component.html',
+    styleUrls: ['./product-list.component.css'],
+    standalone: false
 })
 export class ProductListComponent {
   products = [...products];
 
   share() {
-    window.alert('The product has been shared!');
+    window.alert('Esto es un window.alert emergente que se activa con la funcion share()!');
   }
 
   onNotify() {
-    window.alert('You will be notified when the product goes on sale');
+    window.alert('Te avisa cuando este de oferta, esto es una lert de onNotify()');
   }
 }
 
